+++ shelter/__init__.py
@@ -3,5 +3,5 @@ Simple Python's Tornado wrapper which provides helpers for creating
 a new project, writing management commands, service processes, ...
 """
 
-version_info = (1, 1, 3)
+version_info = (1, 1, 4)
 version = '.'.join(str(i) for i in version_info)